Do not cut unknown word forms from first non-alphabetic character in analysis, test case added
git-svn-id: svn://svn.code.sf.net/p/hfst/code/trunk/hfst3@1518 941e2c2b-deac-454f-805a-451daa25f33c

+++ test/tools/proc-functionality.sh
@@ -71,4 +71,13 @@ fi
 if ! diff test.strings $srcdir/proc-caps-out5.strings ; then
     exit 1
 fi
+
+if ! ../../tools/src/hfst-proc/hfst-proc compounds.hfst.ol < $srcdir/proc-compounds.strings > test.strings ; then
+    echo compound fail:
+    cat test.strings
+    exit 1
+fi
+if ! diff test.strings $srcdir/proc-compounds-out.strings ; then
+    exit 1
+fi
 rm test.strings
